Fix determining the workflows location
Workflows are stored in the `Alfred.alfredrpreferences/workflows`
directory, by default in ~/Library/Application Support/Alfred/`.

The user can override this by setting a sync folder in the app, but if
that was never done, the respective config entry didn't exist, causing
the `link` command of thie CLI to fail.

This chanke makes sure we first search in the override directory, and
only then in the default location.